+++ lib/db.dart
@@ -14,12 +14,15 @@ import 'dart:core';
 import 'dart:core' as core;
 import 'dart:mirrors' as mirrors;
 
+import 'package:meta/meta.dart';
+
 import 'common.dart' show StreamFromPages;
 import 'datastore.dart' as ds;
 import 'service_scope.dart' as ss;
 
 part 'src/db/annotations.dart';
 part 'src/db/db.dart';
+part 'src/db/exceptions.dart';
 part 'src/db/models.dart';
 part 'src/db/model_db.dart';
 part 'src/db/model_db_impl.dart';