+++ lib/pubsub.dart
@@ -7,9 +7,10 @@ library gcloud.pubsub;
 import 'dart:async';
 import 'dart:collection';
 import 'dart:convert';
+import 'dart:io';
 import 'package:http/http.dart' as http;
 
-import 'package:googleapis_beta/pubsub/v1beta2.dart' as pubsub;
+import 'package:googleapis/pubsub/v1.dart' as pubsub;
 
 import 'common.dart';
 import 'service_scope.dart' as ss;
@@ -111,7 +112,7 @@ void registerPubSubService(PubSub pubsub) {
 ///
 abstract class PubSub {
   /// List of required OAuth2 scopes for Pub/Sub operation.
-  static const SCOPES = const [pubsub.PubsubApi.PubsubScope];
+  static const SCOPES = [pubsub.PubsubApi.PubsubScope];
 
   /// Access Pub/Sub using an authenticated client.
   ///
@@ -122,7 +123,12 @@ abstract class PubSub {
   ///
   /// Returs an object providing access to Pub/Sub. The passed-in [client] will
   /// not be closed automatically. The caller is responsible for closing it.
-  factory PubSub(http.Client client, String project) = _PubSubImpl;
+  factory PubSub(http.Client client, String project) {
+    var emulator = Platform.environment['PUBSUB_EMULATOR_HOST'];
+    return emulator == null
+        ? _PubSubImpl(client, project)
+        : _PubSubImpl.rootUrl(client, project, "http://$emulator/");
+  }
 
   /// The name of the project.
   String get project;
@@ -160,7 +166,7 @@ abstract class PubSub {
   ///
   /// Returns a `Future` which completes with a `Page` object holding the
   /// first page. Use the `Page` object to move to the next page of topics.
-  Future<Page<Topic>> pageTopics({int pageSize: 50});
+  Future<Page<Topic>> pageTopics({int pageSize = 50});
 
   /// Create a new subscription named [name] listening on topic [topic].
   ///
@@ -214,7 +220,7 @@ abstract class PubSub {
   /// first page. Use the `Page` object to move to the next page of
   /// subscriptions.
   Future<Page<Subscription>> pageSubscriptions(
-      {String topic, int pageSize: 50});
+      {String topic, int pageSize = 50});
 }
 
 /// A Pub/Sub topic.
@@ -331,7 +337,7 @@ abstract class Subscription {
   ///
   /// If [wait] is `false`, the method will complete the returned `Future`
   /// with `null` if it finds that there are no messages available.
-  Future<PullEvent> pull({bool wait: true});
+  Future<PullEvent> pull({bool wait = true});
 }
 
 /// The content of a Pub/Sub message.