+++ test/db/db_test.dart
@@ -4,7 +4,10 @@
 
 library gcloud.db_test;
 
+import 'dart:mirrors' show reflectClass;
+
 import 'package:gcloud/db.dart';
+import 'package:meta/meta.dart';
 import 'package:test/test.dart';
 
 @Kind()
@@ -13,7 +16,7 @@ class Foobar extends Model {}
 main() {
   group('db', () {
     test('default-partition', () {
-      var db = new DatastoreDB(null);
+      var db = DatastoreDB(null);
 
       // Test defaultPartition
       expect(db.defaultPartition.namespace, isNull);
@@ -30,8 +33,8 @@ main() {
     });
 
     test('non-default-partition', () {
-      var nsDb = new DatastoreDB(null,
-          defaultPartition: new Partition('foobar-namespace'));
+      var nsDb =
+          DatastoreDB(null, defaultPartition: Partition('foobar-namespace'));
 
       // Test defaultPartition
       expect(nsDb.defaultPartition.namespace, 'foobar-namespace');
@@ -46,5 +49,52 @@ main() {
       expect(key.id, 42);
       expect(key.type, equals(Foobar));
     });
+
+    test('hasDefaultConstructor', () {
+      expect(hasDefaultConstructor(Empty), isTrue);
+      expect(hasDefaultConstructor(OnlyNamedConstructor), isFalse);
+      expect(hasDefaultConstructor(DefaultAndNamedConstructor), isTrue);
+      expect(hasDefaultConstructor(RequiredArguments), isFalse);
+      expect(hasDefaultConstructor(OnlyPositionalArguments), isTrue);
+      expect(hasDefaultConstructor(OnlyNamedArguments), isTrue);
+      expect(hasDefaultConstructor(RequiredNamedArguments), isFalse);
+      expect(hasDefaultConstructor(DefaultArgumentValues), isTrue);
+    });
   });
+}
+
+bool hasDefaultConstructor(Type type) =>
+    ModelDBImpl.hasDefaultConstructor(reflectClass(type));
+
+class Empty {
+  const Empty();
+}
+
+class OnlyNamedConstructor {
+  const OnlyNamedConstructor.named();
+}
+
+class DefaultAndNamedConstructor {
+  const DefaultAndNamedConstructor();
+  const DefaultAndNamedConstructor.named();
+}
+
+class RequiredArguments {
+  const RequiredArguments(int arg);
+}
+
+class OnlyPositionalArguments {
+  const OnlyPositionalArguments([int arg, int arg2]);
+}
+
+class OnlyNamedArguments {
+  const OnlyNamedArguments({int arg, int arg2});
+}
+
+class RequiredNamedArguments {
+  const RequiredNamedArguments({int arg1, @required int arg2});
+}
+
+class DefaultArgumentValues {
+  const DefaultArgumentValues([int arg1 = 1, int arg2 = 2]);
 }