+++ test/db_all_e2e_test.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+@Tags(["e2e"])
+
 library gcloud.test.db_all_test;
 
 import 'dart:async';
@@ -19,17 +21,18 @@ import 'db/e2e/metamodel_test_impl.dart' as db_metamodel_test;
 
 Future main() async {
   var scopes = datastore_impl.DatastoreImpl.SCOPES;
-  var now = new DateTime.now().millisecondsSinceEpoch;
-  String namespace = '${Platform.operatingSystem}${now}';
+  var now = DateTime.now().millisecondsSinceEpoch;
+  String namespace = '${Platform.operatingSystem}$now';
 
   datastore_impl.DatastoreImpl datastore;
   db.DatastoreDB datastoreDB;
   Client client;
 
   await withAuthClient(scopes, (String project, httpClient) {
-    datastore = new datastore_impl.DatastoreImpl(httpClient, project);
-    datastoreDB = new db.DatastoreDB(datastore);
+    datastore = datastore_impl.DatastoreImpl(httpClient, project);
+    datastoreDB = db.DatastoreDB(datastore);
     client = httpClient;
+    return null;
   });
 
   tearDownAll(() async {
@@ -45,7 +48,7 @@ Future main() async {
   });
 
   test('sleep-between-test-suites', () {
-    expect(new Future.delayed(const Duration(seconds: 10)), completes);
+    expect(Future.delayed(const Duration(seconds: 10)), completes);
   });
 
   group('datastore_test', () {
@@ -53,7 +56,7 @@ Future main() async {
   });
 
   test('sleep-between-test-suites', () {
-    expect(new Future.delayed(const Duration(seconds: 10)), completes);
+    expect(Future.delayed(const Duration(seconds: 10)), completes);
   });
 
   group('datastore_test', () {